+++ src/lib.rs
@@ -131,6 +131,7 @@ extern crate serde_json;
 pub mod bam;
 pub mod bcf;
 pub mod bgzf;
+pub mod errors;
 pub mod faidx;
 pub mod htslib;
 pub mod tbx;